iwlwifi: fix kdump hang
Patchwork-id: 23841
O-Subject: [RHEL6 PATCH] iwlwifi: fix kdump hang
Bugzilla: 575122
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

BZ#575122

Description (from upstream commit):
To ensure that card is in a sane state during probe we add a reset call.
This change was prompted by users of kdump who was not able to bring up the
wireless driver in the kdump kernel. The problem here was that the primary
kernel, which is not running at the time, left the wireless card up and
running. When the kdump kernel starts it is thus possible to immediately
receive interrupts from firmware after registering interrupt, but without
being ready to deal with interrupts from firmware yet.

Upstream:
commit 4843b5a731b31916d100cfc5ba4d03ae78462ed9
Author: Reinette Chatre <reinette.chatre@intel.com>
Date:   Wed Feb 3 09:38:59 2010 -0800

     iwlwifi: reset card during probe

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=2354705

Testing:
Tested by me on laptop with iwl3945 device.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>