callhome: fix broken proc interface and activate comp ID
Patchwork-id: 23936
O-Subject: [RHEL6 PATCH 1/1] [s390] callhome: fix broken proc interface and
	activate comp ID
Bugzilla: 579482
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
The proc interface of the driver did not accept valid input and
the component ID has not yet been assigned.

strict_strtoul() requires _either_ '\n\0' _or_ '\0' termination.
This was missing and therefore the interface did not recognise
valid input. The solution is to terminate the string properly.
In a case of a kernel panic, ensure that a RHEL Linux instance
on System z can report the crash (panic message) to the correct
PMR queue within the IBM service organization.
(The "callhome" support must be enabled and configured properly).

Bugzilla
--------
BZ 579482
https://bugzilla.redhat.com/show_bug.cgi?id=579482

Upstream status of the patch
----------------------------
The string terminal fix is queued for inclusion into 2.6.34.
The component ID for the correct PMR queue is specific for RHEL.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.