EHCI & UHCI: fix race between root-hub suspend and port resume
Patchwork-id: 23976
O-Subject: [RHEL6 PATCH 13/24] USB: EHCI & UHCI: fix race between root-hub
	suspend and port resume
Bugzilla: 580062
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Alan Stern <stern@rowland.harvard.edu>

commit cec3a53c7fe794237b582e8e77fc0e48465e65ee upstream.

This patch (as1321) fixes a problem with EHCI and UHCI root-hub
suspends: If the suspend occurs while a port is trying to resume, the
resume doesn't finish and simply gets lost.  When remote wakeup is
enabled, this is undesirable behavior.

The patch checks first to see if any port resumes are in progress, and
if they are then it fails the root-hub suspend with -EBUSY.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>