md: fix small irregularity with start_ro module parameter
Patchwork-id: 24020
O-Subject: [RHEL6 PATCH 33/50] md: fix small irregularity with start_ro module
	parameter
Bugzilla: 580063
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: NeilBrown <neilb@suse.de>

commit 0f9552b5dc4fe10da37fa3f4a4ca185d90fa41c9 upstream.

The start_ro modules parameter can be used to force arrays to be
started in 'auto-readonly' in which they are read-only until the first
write.  This ensures that no resync/recovery happens until something
else writes to the device.  This is important for resume-from-disk
off an md array.

However if an array is started 'readonly' (by writing 'readonly' to
the 'array_state' sysfs attribute) we want it to be really 'readonly',
not 'auto-readonly'.

So strengthen the condition to only set auto-readonly if the
array is not already read-only.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>