UBI: initialise update marker
Patchwork-id: 24028
O-Subject: [RHEL6 PATCH 40/50] UBI: initialise update marker
Bugzilla: 580063
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: Peter Horton <zero@colonel-panic.org>

commit ff998793288b49a3b22d929bf8e56362320905ff upstream.

The in kernel copy of a volume's update marker is not initialised from the
volume table. This means that volumes where an update was unfinnished will
not be treated as "forbidden to use". This is basically that the update
functionality was broken.

Signed-off-by: Peter Horton <zero@colonel-panic.org>
Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>