mpt2sas: Reworked scmd->result priority for _scsih_qcmd
Patchwork-id: 24047
O-Subject: [RHEL6 PATCH] mpt2sas: Reworked scmd->result priority for _scsih_qcmd
Bugzilla: 577909
RH-Acked-by: David Milburn <dmilburn@redhat.com>

Aris: this is second part, the first is "mpt2sas: IOs needs to be pause until handles are refreshed for all device after recovery"
which has the same bz.

This is for bz#577909

LSI has found a problem in the mpt2sas driver,
better said this is a fix of the previous fix.
The patch is posted on linux-scsi.
http://marc.info/?l=linux-scsi&m=127073139606404&w=2

LSI:"we added support to set the deleted flag prior to device scan,
then clear the flag for responding devices, leaving the deleted flag only
set for missing devices.  The problem is for internal generated host resets,
IO queues are not blocked at scsi mid layer level.  IO will be continued
sent to driver, and driver  will return SCSI_MLQUEUE_HOST_BUSY.  The problem
is the driver checks for the deleted flag before it checks for the
controller being in reset, so there is a window where the driver would be
returning DID_NO_CONNECT for responding devices.  This occurs during  the
time between calling  _scsih_prep_device_scan, and
_scsih_mark_responding_sas_device & _scsih_mark_responding_raid_device.

Fix the queuecommand entry point so ioc->shost_recovery flag sanity check is
given higher presidence then the device "deleted flag" check.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>