virtio: console: Add support for nonblocking write()s
Patchwork-id: 24089
O-Subject: [RHEL6 PATCH 18/18] virtio: console: Add support for nonblocking
	write()s
Bugzilla: 576241
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

If the host port is not open, a write() should either just return if the
file is opened in non-blocking mode, or block till the host port is
opened.

Also, don't spin till host consumes data for nonblocking ports. For
non-blocking ports, we can do away with the spinning and reclaim the
buffers consumed by the host on the next write call or on the condition
that'll make poll return.

Upstream commit: <posted>
Bugzilla: 576241

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>