nss: add missing .previous statement to asm function
Patchwork-id: 24108
O-Subject: [RHEL6 PATCH 1/1] [s390] nss: add missing .previous statement to
	asm function
Bugzilla: 581521
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
The savesys_ipl_nss() asm function is put into the .init.text section.
However, it is missing a ".previous" statement which would restore the
previous section.  Depending on the compiler version used to built the
kernel, code may or may not be placed into the correct section. If the
code is put into the wrong section, the kernel might panic or fail to
IPL.

To solve this issue, add a the ".previous" statement to return to the
previous section.
section.

Bugzilla
--------
BZ 581521
https://bugzilla.redhat.com/show_bug.cgi?id=581521

Upstream status of the patch
----------------------------
The patch will bwill be upstream as of kernel version 2.6.34
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=187644636ea79aa4bdc04757ebca919c4fd546e0

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.