spec: move the gpg key creation from build to prep
Patchwork-id: 4341
O-Subject: [kernel team] [RHEL6 PATCH] spec: move the gpg key creation from
	build to prep
Bugzilla:
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

When one wants to build the kernel locally, without using RPM, but from a CVS
package/checkout, the gpg keys need to be created manually. Instead of having
an extra script for that, simply move the key creation to %prep section. This
way, 'make prep' or 'rpmbuild -bp' will create the keys automatically.
Tested in a local build and in a scratch brew build to make sure nothing was
broken.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>