raw: add an fsync method
Patchwork-id: 24125
O-Subject: [RHEL6 PATCH v2 1/2] raw: add an fsync method
Bugzilla: 579781
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

Hi,

O_SYNC writes and fsync on a raw device will fail with -EINVAL.  Anton
Blanchard posted a patch to fix this upstream, describing the problem
thusly:

  Commit 148f948ba877f4d3cdef036b1ff6d9f68986706a (vfs: Introduce new
  helpers for syncing after writing to O_SYNC file or IS_SYNC inode)
  broke the raw driver.

  We now call through generic_file_aio_write -> generic_write_sync ->
  vfs_fsync_range. vfs_fsync_range has:

        if (!fop || !fop->fsync) {
                ret = -EINVAL;
                goto out;
        }
  But drivers/char/raw.c doesn't set an fsync method.

  We have two options: fix it or remove the raw driver completely. I'm happy
  to do either, the fact this has been broken for so long suggests it is
  rarely used.

  The patch below adds an fsync method to the raw driver. My knowledge of the
  block layer is pretty sketchy so this could do with a once over.

  If we instead decide to remove the raw driver, this patch might still be
  useful as a backport to 2.6.33 and 2.6.32.

  Signed-off-by: Anton Blanchard <anton@samba.org>

Given that databases often open the raw device, we need to keep the raw
driver and fix this.  I tested the provided patch on a RHEL 6 kernel and
it works for me.

This addresses bug 579781.  Comments, as always, are appreciated.

Cheers,
Jeff
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>