rename block_fsync() to blkdev_fsync()
Patchwork-id: 24123
O-Subject: [RHEL6 PATCH v2 2/2] vfs: rename block_fsync() to blkdev_fsync()
Bugzilla: 579781
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

Hi,

This is a follow-on patch to adding an fsync routine for raw devices.
As noted below, hch asked for the block_fsync function to be renamed
for consistency.

  From: Andrew Morton <akpm@linux-foundation.org>

  Requested by hch, for consistency now it is exported.

  Signed-off-by: Andrew Morton <akpm@linux-foundation.org>

This is related to bug 579781.  Comments, as always, are appreciated.

Cheers,
Jeff
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>