b43: Rewrite DMA Tx status handling sanity checks
Patchwork-id: 24137
O-Subject: [RHEL6 PATCH] b43: Rewrite DMA Tx status handling sanity checks
Bugzilla: 574533
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Back-port of the following upstream commit...

commit 07681e211d736ba2394ab7f29f77e93adecd22c5
Author: Michael Buesch <mb@bu3sch.de>
Date:   Thu Nov 19 22:24:29 2009 +0100

    b43: Rewrite DMA Tx status handling sanity checks

    This rewrites the error handling policies in the TX status handler.
    It tries to be error-tolerant as in "try hard to not crash the machine".
    It won't recover from errors (that are bugs in the firmware or driver),
    because that's impossible. However, it will return a more or less useful
    error message and bail out. It also tries hard to use rate-limited messages
    to not flood the syslog in case of a failure.

    Signed-off-by: Michael Buesch <mb@bu3sch.de>
    Signed-off-by: John W. Linville <linville@tuxdriver.com>

Without this patch, it is possible (particular if using the open
source firmware) for the b43 driver to crash the machine as a result
of a firmware error.

BZ574533

I have received no reports of this on RHEL6, but I'm sure the problem
exists there.  This patch was tested w/ good results on upstream and
Fedora kernels in the 2.6.32 era.  At worst, I believe it does no
harm at very little cost.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>