build: fix a false warning of linux-kernel-test.patch not being in sync
Patchwork-id: 4343
O-Subject: [kernel team] [RHEL6 PATCH] build: fix a false warning
	of	linux-kernel-test.patch not being in sync
Bugzilla:

Most of the rh- targets end up depending on sources-rh target and it checks
if the current changes on the git tree match linux-kernel-test.patch contents.
This is used to print a warning in the case they're not in sync, to remind you
to run 'make rh-test-patch' to refresh linux-kernel-test.patch. Currently the
check doesn't filter out redhat/ changes like 'make rh-test-patch', so it'll
falsely warn about linux-kernel-test.patch not being in sync. This patch solves
the issue. Reported and tested by Don Zickus.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>