gss_krb5: Use confounder length in wrap code
Patchwork-id: 24195
O-Subject: [RHEL6 PATCH 20/22] gss_krb5: Use confounder length in wrap code
Bugzilla: 498317
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Ric Wheeler <rwheeler@redhat.com>

From: Kevin Coffman <kwc@citi.umich.edu>

All encryption types use a confounder at the beginning of the
wrap token.  In all encryption types except arcfour-hmac, the
confounder is the same as the blocksize.  arcfour-hmac has a
blocksize of one, but uses an eight byte confounder.

Add an entry to the crypto framework definitions for the
confounder length and change the wrap/unwrap code to use
the confounder length rather than assuming it is always
the blocksize.

Signed-off-by: Kevin Coffman <kwc@citi.umich.edu>
Signed-off-by: Steve Dickson <steved@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>