rsize and wsize settings ignored on v4 mounts
Patchwork-id: 24197
O-Subject: [RHEL6] [PATCH] Bug 582697 - NFS: rsize and wsize settings ignored
	on v4 mounts
Bugzilla: 582697
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

This is a pretty obvious one...

The bz is: https://bugzilla.redhat.com/show_bug.cgi?id=582697

steved.

NFSv4 mounts ignore the rsize and wsize mount options, and always use
the default transfer size for both.  This seems to be because all
NFSv4 mounts are now cloned, and the cloning logic doesn't copy the
rsize and wsize settings from the parent nfs_server.

I tested Fedora's 2.6.32.11-99 and it seems to have this problem as
well, so I'm guessing that .33, .32, and perhaps older kernels have
this issue as well.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Cc: Stable <stable@kernel.org>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>