modsign: fix forced module re-signing bug
Patchwork-id: 24204
O-Subject: [RHEL6 PATCH] Fix forced module re-signing bug
Bugzilla: 578916
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

Currently the kernel build re-signs all the modules every time you do a build,
even if it doesn't need to [BZ 578916].

This is because one of the intermediate targets in the production of a module
(the unsigned module, foo.ko.unsigned) wasn't listed in the $(targets) make
variable.  Instead, the final module name (foo.ko) was listed twice.

This is fixed by listing the correct target.  After that, only changed modules
are re-signed.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>