re-export page_is_ram() for crash module
Patchwork-id: 24218
O-Subject: [RHEL6 PATCH]: re-export page_is_ram() for crash module
Bugzilla: 583032
RH-Acked-by: Eric Paris <eparis@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

When I un-exported page_is_ram it was based on the assumption that the Kconfig
listed crash as a bool.  It turns out that I was looking at CRASH_DUMP and not
CRASH -- crash *can* be built as a module.

Re-export page_is_ram() that was removed in BZ 578834.

Successfully compiled by me.

NOT UPSTREAM -- crash is not upstream so this is an RH only patch.

Resolves BZ 583032.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>