tun: orphan an skb on tx
Patchwork-id: 24284
O-Subject: [RHEL6.0 PATCH] tun: orphan an skb on tx
Bugzilla: 584428
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

BZ#584428
https://bugzilla.redhat.com/show_bug.cgi?id=584428

Description
The following situation was observed in the field:
tap1 sends packets, tap2 does not consume them, as a result
tap1 can not be closed. This happens because
tun/tap devices can hang on to skbs undefinitely.

As noted by Herbert, possible solutions include a timeout followed by a
copy/change of ownership of the skb, or always copying/changing
ownership if we're going into a hostile device.

This patch implements the second approach.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Tested-by: Yan Vugenfirer <yvugenfi@redhat.com>
Acked-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>
Tested-by: Jan Kiszka <jan.kiszka@siemens.com>

Upstream status
This patch has been merged for 2.6.34, commit
0110d6f22f392f976e84ab49da1b42f85b64a3c5

Test status
The patch was tested in-house on x86_64, an upstream version of this
patch was tested by jan.kiszka@siemens.com at customer's lab.

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2393193

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>