show per-process swap usage via procfs
Patchwork-id: 24287
O-Subject: Re: [RHEL6 patch] show per-process swap usage via procfs
Bugzilla: 546533
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

Fujitsu posted a patch that keeps track of the swap space usage on a
per-process basis and make is visible via /proc/<pid>/status.  This
patch was posted to lkml and linux-mm last fall.  The patch went through
several review and update iterations and was finally ACK'd.  I dont know
exactly what the state of this patch is but it will probably eventually
make it into the upstream kernel.  The patch provides useful debugging
information, has be requested by customers.

The patch adds the VmSwap: row of data to the /proc/<pid#>/status
output:

	FDSize: 256
	Groups: 0 1 2 3 4 6 10
	VmPeak: 33558216 kB
	VmSize: 33558216 kB
	VmLck:         0 kB
	VmHWM:  15916920 kB
	VmRSS:  15876632 kB
	VmData: 33554480 kB
	VmStk:        84 kB
	VmExe:         4 kB
	VmLib:      1568 kB
	VmPTE:     32804 kB
>>>>>>>>VmSwap:   900808 kB
		Threads:        1
	SigQ:   1/128638
	SigPnd: 0000000000000000

The upstream patch submission, discussion and ACKs are located on
http://lwn.net/Articles/360307/

The attached RHEL6 packport of this patch addresses BZ546533.

v2:
Good point, attached is an updated patch that includes the upstream
changes made to Documentation/filesystems/proc.txt(minimal as they are).

v3:
Attached is an updated, tested and verified patch that makes the
suggestions and aligns with the latest upstream kernel.

Larry

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>