Specify REGULATOR_CHANGE_STATUS for WM835x LED constraints
Patchwork-id: 24317
O-Subject: [RHEL6 PATCH 20/47] regulator: Specify REGULATOR_CHANGE_STATUS for
	WM835x LED constraints
Bugzilla: 580064
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: Mark Brown <broonie@opensource.wolfsonmicro.com>

commit a2fad9bf26a1d44a8d31a5c4528108a2b9f468ab upstream.

The WM8350 LED driver needs to be able to enable and disable the
regulators it is using. Previously the core wasn't properly enforcing
status change constraints so the driver was able to function but this
has always been intended to be required.

Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>
Signed-off-by: Liam Girdwood <lrg@slimlogic.co.uk>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>