adt7462: fix wrong ADT7462_VOLT_COUNT
Patchwork-id: 24348
O-Subject: [RHEL6 PATCH 4/50] hwmon: (adt7462) Wrong ADT7462_VOLT_COUNT
Bugzilla: 580065
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

From: Ray Copeland <ray.copeland@aprius.com>

commit 85f8d3e5faea8bd36c3e5196f8334f7db45e19b2 upstream.

The #define ADT7462_VOLT_COUNT is wrong, it should be 13 not 12. All the
for loops that use this as a limit count are of the typical form, "for
(n = 0; n < ADT7462_VOLT_COUNT; n++)", so to loop through all voltages
w/o missing the last one it is necessary for the count to be one greater
than it is.  (Specifically, you will miss the +1.5V 3GPIO input with count
= 12 vs. 13.)

Signed-off-by: Ray Copeland <ray.copeland@aprius.com>
Acked-by: "Darrick J. Wong" <djwong@us.ibm.com>
Signed-off-by: Jean Delvare <khali@linux-fr.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>