b43/b43legacy: Wake queues in wireless_core_start
Patchwork-id: 24459
O-Subject: [RHEL6 PATCH 64/93] b43/b43legacy: Wake queues in wireless_core_start
Bugzilla: 580067
RH-Acked-by: John Linville <linville@redhat.com>

From: Larry Finger <Larry.Finger@lwfinger.net>

commit 0866b03c7d7dee8a34ffa527ecda426c0f405518 upstream.

If b43 or b43legacy are deauthenticated or disconnected, there is a
possibility that a reconnection is tried with the queues stopped in
mac80211. To prevent this, start the queues before setting
STAT_INITIALIZED.

In b43, a similar change has been in place (twice) in the
wireless_core_init() routine. Remove the duplicate and add similar
code to b43legacy.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: John W. Linville <linville@tuxdriver.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>