ring-buffer: Move disabled check into preempt disable section
Patchwork-id: 24499
O-Subject: [RHEL6 PATCH 8/86] ring-buffer: Move disabled check into preempt
	disable section
Bugzilla: 584757
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: Lai Jiangshan <laijs@cn.fujitsu.com>

commit 52fbe9cde7fdb5c6fac196d7ebd2d92d05ef3cd4 upstream.

The ring buffer resizing and resetting relies on a schedule RCU
action. The buffers are disabled, a synchronize_sched() is called
and then the resize or reset takes place.

But this only works if the disabling of the buffers are within the
preempt disabled section, otherwise a window exists that the buffers
can be written to while a reset or resize takes place.

Reported-by: Li Zefan <lizf@cn.fujitsu.com>
Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>
LKML-Reference: <4B949E43.2010906@cn.fujitsu.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>