DVB: em28xx-dvb: fix memleak in dvb_fini()
Patchwork-id: 24503
O-Subject: [RHEL6 PATCH 16/86] V4L/DVB (13961): em28xx-dvb: fix memleak in
	dvb_fini()
Bugzilla: 584757
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Mauro Carvalho Chehab <mchehab@redhat.com>

From: Francesco Lavra <francescolavra@interfree.it>

commit 19f48cb105b7fa18d0dcab435919a3a29b7a7c4c upstream.

this patch fixes a memory leak which occurs when an em28xx card with DVB
extension is unplugged or its DVB extension driver is unloaded. In
dvb_fini(), dev->dvb must be freed before being set to NULL, as is done
in dvb_init() in case of error.
Note that this bug is also present in the latest stable kernel release.

Signed-off-by: Francesco Lavra <francescolavra@interfree.it>
Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>