coretemp: Add missing newline to dev_warn() message
Patchwork-id: 24567
O-Subject: [RHEL6 PATCH 73/86] hwmon: (coretemp) Add missing newline to
	dev_warn() message
Bugzilla: 584757
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

From: Dean Nelson <dnelson@redhat.com>

commit 4d7a5644e4adfafe76c2bd8ee168e3f3b5dae3a8 upstream.

Add missing newline to dev_warn() message string. This is more of an issue
with older kernels that don't automatically add a newline if it was missing
from the end of the previous line.

Signed-off-by: Dean Nelson <dnelson@redhat.com>
Signed-off-by: Jean Delvare <khali@linux-fr.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>